Fixed UnturnedPlayerManager.Ban()
It used `Provider.ban()`, which doesn't actually ban the player, only disconnects them from the server as if they were banned (Changed to `SteamBlacklist.ban()` which was correctly used for when the player was offline, but not when online).
Also changed `var player = ((UnturnedUser)target).Player;` for `if (target is UnturnedUser user)`.
Removed as well the second time it casted Target to UnturnedUser when player was already defined.